1.connected user to trips, now each created trip has an owner; 2.delete trip works, but the trip id remains in user trips array; 3.Created separate pages for public, planned and completed trips

+++ src/utils/api.js
@@ -25,17 +25,17 @@ export function getTrip(tripId) {
     return client.get(`/trips/${tripId}`)
 }
 
-export function deleteTrip(tripId) {
-  return client.delete(`/trips/${tripId}`)
+export function deleteTrip(tripId, ownerId) {
+  return client.delete(`/trips/${ownerId}/${tripId}`)
 }
 
-export function createTrip(data) {
-  return client.post("/trips", data, {
-    headers: {
-      "Content-Type": "application/json"
-    }
-  })
-}
+// export function createTrip(data) {
+//   return client.post("/trips", data, {
+//     headers: {
+//       "Content-Type": "application/json"
+//     }
+//   })
+// }
 
 export function updateTrip(tripId, data) {
   return client.put(`/trips/${tripId}`, JSON.stringify(data), {
